discuss: Use target=_blank for all cfbot links
This change starts opening all the CFbot related links in a new tab.

All these links are to external websites, some people on the internet
consider it best practice to open links to external sites in a separate
tab. Some other people on the internet strongly disagree with that
belief. [1]

We should have some discussion about this, to discuss what behaviour we
want for these links (maybe it even depends on the link).